requirements.txt
aniso8601==9.0.1
bcrypt==4.1.2
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
Flask==3.0.2
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-RESTful==0.3.10
Flask-WTF==1.2.1
idna==3.7
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
oracledb==2.1.1
pillow==10.3.0
pycparser==2.22
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
six==1.16.0
urllib3==2.2.1
Werkzeug==3.0.2
WTForms==3.1.2
gunicorn==21.2.0
